move ucx conf to solitary file
